functions, types and tests migration
